Combine existing contractor metadata test
